json

+++ RPGM/Chanomhub-RPGM/main.go
@@ -11,23 +11,23 @@ import (
 )
 
 type EnemyData struct {
-	ID             int        `json:"id"`
-	Animation1Hue  int        `json:"animation1Hue"`
-	Animation1Name string     `json:"animation1Name"`
-	Animation2Hue  int        `json:"animation2Hue"`
-	Animation2Name string     `json:"animation2Name"`
-	Frames         [][][]int  `json:"frames"`
-	Name           string     `json:"name"`
-	Position       int        `json:"position"`
-	Timings        []Timings  `json:"timings"`
+	ID             int           `json:"id"`
+	Animation1Hue  int           `json:"animation1Hue"`
+	Animation1Name string        `json:"animation1Name"`
+	Animation2Hue  int           `json:"animation2Hue"`
+	Animation2Name string        `json:"animation2Name"`
+	Frames         [][][]float64 `json:"frames"`
+	Name           string        `json:"name"`
+	Position       int           `json:"position"`
+	Timings        []Timings     `json:"timings"`
 }
 
 type Timings struct {
-	FlashColor   []int      `json:"flashColor"`
-	FlashDuration int       `json:"flashDuration"`
-	FlashScope    int        `json:"flashScope"`
-	Frame         int        `json:"frame"`
-	Se            Se         `json:"se"`
+	FlashColor    []int `json:"flashColor"`
+	FlashDuration int   `json:"flashDuration"`
+	FlashScope    int   `json:"flashScope"`
+	Frame         int   `json:"frame"`
+	Se            Se    `json:"se"`
 }
 
 type Se struct {
@@ -36,7 +36,6 @@ type Se struct {
 	Pitch  int    `json:"pitch"`
 	Volume int    `json:"volume"`
 }
-
 
 func main() {
 	// Command-line flags
